AUS-4285 Changed to use countMax.

+++ src/main/java/org/auscope/nvcl/server/service/TSGModJobProcessor.java
@@ -41,7 +41,7 @@ public class TSGModJobProcessor  extends IJobProcessor{
     private String tsgScript;
     private String dataPath;
     private String wvLogname = "Reflectance";
-    private double countSumMax = 0.0;
+    private Integer countSumMax = 0;
     
 	    
     /**
